% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultOutputRequest.R
\name{defaultOutput}
\alias{defaultOutput}
\title{Make a default list of output requests.}
\usage{
defaultOutput(gangstas)
}
\arguments{
\item{gangstas}{The list of gangstas in the model.}
}
\value{
A list of output requests that can be passed to \code{gangstaRapper()}
}
\description{
Make a default list of output requests.
}
