% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeGangstaModel.R
\name{executeGangstaModel}
\alias{executeGangstaModel}
\title{Execute one iteration of a GANGSTA Model}
\usage{
executeGangstaModel(rapper = parent.frame())
}
\arguments{
\item{rapper}{The gangstaRapper environment containing the lpModel and all 
other variables and functions needed to update and solve the lpModel. 
The default rapper argument for this function is parent.frame(),
which references the global environment.}
}
\value{
A list of requested model output for a single time step.
}
\description{
The executeRapper function will iterately update driving values
and call this function, which updates the lp model, runs the model for one time step,
and reports results from the model run.
}
