% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gangstaRapper.R, R/gangstaRapperPackage.R
\docType{package}
\name{gangstaRapper}
\alias{gangstaRapper}
\title{Make a gangstaRapper}
\usage{
gangstaRapper(
  execute = executeGangstaModel,
  drivingValues = NULL,
  leakInDF = NULL,
  lpModelFilePath,
  gangstas,
  updates = list(),
  outputRequest = defaultOutput(gangstas),
  isotopesToTrack = NA,
  initialAtomicFractions = NA,
  isotopeStandards = NA,
  ...,
  initValues = list()
)
}
\arguments{
\item{execute}{A function that gets called iteratively, i.e. once per time step. 
The default is \code{executeGangstaModel}, which updates and then solves the lpModel
and returns output specified in the outputRequest.}

\item{drivingValues}{A \code{data.frame} with columns named for driving
variables required by the function associated with the \code{execute} argument.
Each row represents the values of the driving variables for one timestep.
The \code{row.names} must be equal to \code{0:nTimeSteps} where nTimeSteps
is the number of timesteps to execute the simulation.}

\item{leakInDF}{A data frame where each column is named according to the convention
"add.to.<lpSolveVariableName>" and values in each row represent the amount added to the 
 variable associated with the column during a single timestep.}

\item{lpModelFilePath}{the file path for the .lp model as a character string}

\item{gangstas}{The list of gangsta objects used to generate the model.}

\item{updates}{placeholder}

\item{outputRequest}{A list of output requests. If this argument is not supplied, 
a default set of outputs will be generated.}

\item{isotopesToTrack}{A named list of numeric vectors where names are elements and vectors
contain the atomic weight of the isotope to track}

\item{initialAtomicFractions}{A named list of named numeric vectors where the
names of the list are names of pools, and numeric vectors contain initial
atomic fractions and are named according to the atomic weight of the isotope}

\item{...}{Named arguments describing model parameters or initial condition
for state variables where the name of the argument represents the variable
name that will store the state variable or parameter value within the rapper environment.}

\item{initValues}{An alternative way to specify parameters and initial values
of state variables using a named list. May be used in conjunction with
\code{...} arguments.}
}
\value{
An object of class gangstaRapper.
}
\description{
Creates a gangstaRapper object. The model is iterated when the 'executeRapper'
function is called with the gangstaRapper object as its argument.

The Gangsta Rapper package facilitates iteration of gangsta generated models.
}
\section{This is a test}{
 test
}

