% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput.R
\name{getOutput}
\alias{getOutput}
\title{Get requested model output from an lp Model.}
\usage{
getOutput(outputRequest, rapper)
}
\arguments{
\item{outputRequest}{A data frame, character vector, or expression vector.}
}
\value{
A data frame or vector of requested model output for a single time step.
}
\description{
Get requested model output from an lp Model.
}
