% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFunctions.R
\name{leakIn}
\alias{leakIn}
\title{Leak in a compound.}
\usage{
leakIn(ID)
}
\arguments{
\item{ID}{The name of the variable  initial molecules of compound.}
}
\description{
An "update" function that leaks in compounds by adding
an amount to the final amount from the previous time step. The amount is 
specified in the rapper environment as "add.to.<compoundName> and gets 
updated with each iteration according the user-specified leak-in
data frame.
}
