% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeGangstaModel.R
\name{updateIsotopes}
\alias{updateIsotopes}
\title{Calculate new isotopic compositions resulting from transfers and update variables.}
\usage{
updateIsotopes(rapper = parent.frame())
}
\description{
Calculates the new isotopic composition of compound pools based
 on the amount of compound transferred to and from the pool, the isotopic
 compositions of "from" pools and leak ins, and the isotopic composition of the compound
 pool at the beginning of the time step.
}
