% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gangstaRapper.R
\name{gangstaRapper}
\alias{gangstaRapper}
\title{Make a gangstaRapper}
\usage{
gangstaRapper(execute = executeGangstaModel, drivingValues = NULL,
  leakInDF = NULL, lpModelFilePath, gangstas, updates = list(),
  outputRequest = NULL, ..., initValues = list())
}
\arguments{
\item{leakInDF}{A data frame where each column is named according to the convention
"add.to.<lpSolveVariableName>" and values in each row represent the amount added to the 
 variable associated with the column during a single timestep.}

\item{lpModelFilePath}{the file path for the .lp model as a character string}

\item{gangstas}{The list of gangsta objects used to generate the model.}

\item{updates}{}

\item{outputRequest}{A list of output requests. If this argument is not supplied, 
a default set of outputs will be generated.}

\item{...}{Named arguments describing model parameters or initial condition
for state variables where the name of the argument represents the variable
name that will store the state variable or parameter value within the rapper environment.}

\item{initValues}{An alternative way to specify parameters and initial values
of state variables using a named list. May be used in conjunction with
\code{...} arguments.}
}
\value{
An object of class gangstaRapper.
}
\description{
Creates a gangstaRapper object. The model is iterated when the 'executeRapper'
function is called with the gangstaRapper object as its argument.
}
